mar-backend/trip-service/requirements.txt
Django>=2.2.7.0
djangorestframework>=3.10.3
django-cors-headers>=3.5.0
django_crontab
djangorestframework-simplejwt>=5.2.2
Pillow
argon2-cffi
gunicorn==20.1.0
psycopg2-binary
mysqlclient 